Modify tc2dracor.xq to transform only a single document posted to it
This moves the control of looping over the source files back to the bash
script. Under macOS both the docker containers as well as podman pod on
a virtual Fedora ran into memory problems when processing the full set
of 1449 files.

Taking the loop out of the XQuery alleviates these problems and
significantly speeds up the entire transformation process. In fact, the
fastest run in under 4min I now get with a single instance of an eXistDB
running directly on macOS.