Refactor transformation xquery
The xquery now only deals with the transformation only. The validation
part will be moved to a separate script.

It iterates over a collection of source documents, preprocesses each
document, constructs a new TEI document and saves it directly to a
directory in the local file system without storing the result in the
database.