Update README
This is supposed to unclutter our list of changes by removing some of
the more particular ones and adding a few more recent ones.

I also simplified the TODO section by referring to the GitHub issues,
where I created two new ones. The error mentioned in this section hasn't
occurred in recent transformations, so I think we can drop that one for
now.