Revise the transformation of speaker IDs
We now recognize the underscore as an additional separator of individual speaker IDs
when multiple speakers occur in an sp. This requires fixing the the use of underscores
in other cases on the dracor branch.

We also use a more sophisticated approach to transform IDs starting with a number, which we now move to the end of the ID.